pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "graphlens"
version = "0.1.0"
description = "Starter template for a Python RAG and vector database project."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.116.0",
  "openai-whisper",
  "python-dotenv>=1.0.1",
  "pydantic>=2.10.0",
  "pyyaml>=6.0.2",
  "streamlit>=1.42.0",
  "uvicorn>=0.35.0",
  "yt-dlp",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
